Moved Twitter API functions to helper file
